Fix "more runs" button repeating the oldest run
Relying on timestamps to paginate the results is not accurate since they
are int64 number of microseconds converted into double number of seconds,
and so we ended up selecting the oldest entry again despite we were
asking for entries strictly before.

Anyway, now we paginate using the id of the oldest top_run of the list.